+++ cmd/babylond/cmd/cmd_test.go
@@ -8,16 +8,19 @@ import (
 
 	"github.com/babylonlabs-io/babylon/v2/app"
 	"github.com/babylonlabs-io/babylon/v2/cmd/babylond/cmd"
+	"github.com/cosmos/cosmos-sdk/client/flags"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 	"github.com/cosmos/cosmos-sdk/x/genutil/client/cli"
 )
 
 func TestInitCmd(t *testing.T) {
+	tmpDir := t.TempDir()
 	rootCmd := cmd.NewRootCmd()
 	rootCmd.SetArgs([]string{
 		"init",     // Test the init cmd
 		"app-test", // Moniker
 		fmt.Sprintf("--%s=%s", cli.FlagOverwrite, "true"), // Overwrite genesis.json, in case it already exists
+		fmt.Sprintf("--%s=%s", flags.FlagHome, tmpDir),
 		"--no-bls-password",
 	})