+++ test/replay/driver.go
@@ -48,6 +48,7 @@ import (
 	xauthsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
+	v1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	gogoprotoio "github.com/cosmos/gogoproto/io"
 	"github.com/otiai10/copy"
 	"github.com/stretchr/testify/require"
@@ -78,6 +79,7 @@ const (
 	defaultGasLimit = 750000
 	defaultFee      = 500000
 	epochLength     = 10
+	blkTime         = time.Second * 5
 )
 
 var (
@@ -143,6 +145,14 @@ type BabylonAppDriver struct {
 	CurrentTime      time.Time
 }
 
+// NewBabylonAppDriverTmpDir initializes Babylon driver for block creation with
+// temporary directories
+func NewBabylonAppDriverTmpDir(r *rand.Rand, t *testing.T) *BabylonAppDriver {
+	driverTempDir := t.TempDir()
+	replayerTempDir := t.TempDir()
+	return NewBabylonAppDriver(r, t, driverTempDir, replayerTempDir)
+}
+
 // Inititializes Babylon driver for block creation
 func NewBabylonAppDriver(
 	r *rand.Rand,
@@ -150,12 +160,14 @@ func NewBabylonAppDriver(
 	dir string,
 	copyDir string,
 ) *BabylonAppDriver {
+	expeditedVotingPeriod := blkTime + time.Second
+
 	chain, err := initialization.InitChain(
 		chainID,
 		dir,
 		[]*initialization.NodeConfig{validatorConfig},
-		3*time.Minute,
-		1*time.Minute,
+		expeditedVotingPeriod*2, // voting period
+		expeditedVotingPeriod,   // expedited
 		1,
 		[]*btclighttypes.BTCHeaderInfo{},
 	)
@@ -270,6 +282,10 @@ func NewBabylonAppDriver(
 		// initiate time to current time
 		CurrentTime: time.Now(),
 	}
+}
+
+func (d *BabylonAppDriver) Ctx() sdk.Context {
+	return d.GetContextForLastFinalizedBlock()
 }
 
 func (d *BabylonAppDriver) GetLastFinalizedBlock() *FinalizedBlock {
@@ -491,7 +507,7 @@ func (d *BabylonAppDriver) GenerateNewBlock() *abci.ResponseFinalizeBlock {
 	// ApplyBlock
 	// add slepp to avoid zero duration for minting
 	// Simulate 5s block time
-	newTime := d.CurrentTime.Add(5 * time.Second)
+	newTime := d.CurrentTime.Add(blkTime)
 	extCommitSig := cmttypes.ExtendedCommitSig{
 		CommitSig: cmttypes.CommitSig{
 			BlockIDFlag:      cmttypes.BlockIDFlagCommit,
@@ -753,7 +769,8 @@ func (d *BabylonAppDriver) WaitTillAllFpsJailed(t *testing.T) {
 // execution was successful. It assumes that there will only be one tx in the block.
 func (d *BabylonAppDriver) SendTxWithMsgsFromDriverAccount(
 	t *testing.T,
-	msgs ...sdk.Msg) {
+	msgs ...sdk.Msg,
+) {
 	d.SendTxWithMessagesSuccess(
 		t,
 		d.SenderInfo,
@@ -901,5 +918,34 @@ func (d *BabylonAppDriver) CreateCovenantSender() *CovenantSender {
 		t:   d.t,
 		d:   d,
 		app: d.App,
+	}
+}
+
+func (d *BabylonAppDriver) GovPropAndVote(msgInGovProp sdk.Msg) (lastPropId uint64) {
+	msgToSend := d.NewGovProp(msgInGovProp)
+	d.SendTxWithMsgsFromDriverAccount(d.t, msgToSend)
+
+	props := d.GovProposals()
+	lastPropId = props[len(props)-1].Id
+
+	d.GovVote(lastPropId)
+	return lastPropId
+}
+
+func (d *BabylonAppDriver) GovPropWaitPass(msgInGovProp sdk.Msg) {
+	propId := d.GovPropAndVote(msgInGovProp)
+
+	for {
+		prop := d.GovProposal(propId)
+
+		if prop.Status == v1.ProposalStatus_PROPOSAL_STATUS_FAILED {
+			d.t.Errorf("prop %d failed due to: %s", propId, prop.FailedReason)
+		}
+
+		if prop.Status == v1.ProposalStatus_PROPOSAL_STATUS_PASSED {
+			break
+		}
+
+		d.GenerateNewBlock()
 	}
 }